+++ src/test/unit/plugin_tests/structured_headings.js
@@ -188,7 +188,7 @@ if (WYMeditor.STRUCTURED_HEADINGS_POLYFILL_REQUIRED) {
         expect(1);
         var wymeditor = jQuery.wymeditors(0);
 
-        wymeditor.html(startHeadings);
+        wymeditor._html(startHeadings);
         equals(getHtmlAfterKeyup(wymeditor), expectedHeadings,
                "Heading numbering properly added on keyup");
     });
@@ -198,25 +198,25 @@ if (WYMeditor.STRUCTURED_HEADINGS_POLYFILL_REQUIRED) {
         var wymeditor = jQuery.wymeditors(0);
 
         // Add correct headings to editor first
-        wymeditor.html(expectedHeadings);
+        wymeditor._html(expectedHeadings);
         equals(getHtmlAfterKeyup(wymeditor), expectedHeadings,
                "Heading numbering unchanged on keyup if already correct");
 
         // Break a correct top level heading by replacing it with a heading
         // with edited numbering
-        wymeditor.html(editedNumberingTopLevel);
+        wymeditor._html(editedNumberingTopLevel);
         equals(getHtmlAfterKeyup(wymeditor), expectedHeadings,
                "Top level heading numbering fixed on keyup");
 
         // Break a correct sublevel heading by replacing it with a heading with
         // edited numbering
-        wymeditor.html(editedNumberingSubLevel);
+        wymeditor._html(editedNumberingSubLevel);
         equals(getHtmlAfterKeyup(wymeditor), expectedHeadings,
                "Sublevel heading numbering fixed on keyup");
 
         // Break a correct subsublevel heading by replacing it with a heading
         // with edited numbering
-        wymeditor.html(editedNumberingSubSubLevel);
+        wymeditor._html(editedNumberingSubSubLevel);
         equals(getHtmlAfterKeyup(wymeditor), expectedHeadings,
                "Subsublevel heading numbering fixed on keyup");
     });